+++ craft_store/__init__.py
@@ -21,9 +21,10 @@ __version__ = "3.0.0"
 
 
 from . import creds, endpoints, errors, models
+from ._httpx_auth import CandidAuth, DeveloperTokenAuth
+from .publisher import PublisherGateway
 from .auth import Auth
 from .base_client import BaseClient
-from .developer_token_auth import DeveloperTokenAuth
 from .http_client import HTTPClient
 from .store_client import StoreClient
 from .ubuntu_one_store_client import UbuntuOneStoreClient
@@ -33,8 +34,10 @@ __all__ = [
     "endpoints",
     "errors",
     "models",
+    "PublisherGateway",
     "Auth",
     "BaseClient",
+    "CandidAuth",
     "HTTPClient",
     "StoreClient",
     "UbuntuOneStoreClient",